# api/openapi/openapi.yaml
openapi: 3.1.0
info:
  title: Amadya Core API
  version: 0.2.0
  license:
    name: Proprietary
    identifier: LicenseRef-Proprietary
  description: |
    Contract-first API for the Amadya HoReCa platform (see docs/adr/0002-contract-first-openapi.md).
    Localized content is resolved using the `Accept-Language` header (`ro`, `en`; default from settings).
    Errors use RFC 9457 problem+json with a stable `code`.

    Tags `streams` (Server-Sent Events) and `webhooks` (raw payloads) are implemented by hand on the server;
    every other tag is implemented through the generated interfaces.
servers:
  - url: /api/v1
...
security:
  - bearerAuth: []
paths:
...
  /admin/units:
    get:
      tags: [inventory]
      operationId: listMeasureUnits
      summary: Units of measure (reference unit per dimension + conversion factors)
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema: { type: array, items: { $ref: "#/components/schemas/MeasureUnit" } }
        "403": { $ref: "#/components/responses/Problem" }
    post:
      tags: [inventory]
      operationId: createMeasureUnit
      summary: Create a unit (factor relative to the dimension's reference unit)
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/MeasureUnitRequest" }
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema: { $ref: "#/components/schemas/MeasureUnit" }
        "409": { $ref: "#/components/responses/Problem" }
  /admin/units/{id}:
    put:
      tags: [inventory]
      operationId: updateMeasureUnit
      summary: Update a unit; mapping an UNMAPPED unit (from an invoice) sets its dimension and factor
      parameters: [{ $ref: "#/components/parameters/Id" }]
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/MeasureUnitRequest" }
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema: { $ref: "#/components/schemas/MeasureUnit" }
        "409": { $ref: "#/components/responses/Problem" }
  /admin/warehouses:
    get:
      tags: [inventory]
      operationId: listWarehouses
      summary: List warehouses (gestiuni)
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema: { type: array, items: { $ref: "#/components/schemas/Warehouse" } }
        "403": { $ref: "#/components/responses/Problem" }
    post:
      tags: [inventory]
      operationId: createWarehouse
      summary: Create a warehouse
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/WarehouseRequest" }
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Warehouse" }
        "409": { $ref: "#/components/responses/Problem" }
  /admin/warehouses/{id}:
    put:
      tags: [inventory]
      operationId: updateWarehouse
      summary: Update a warehouse
      parameters: [{ $ref: "#/components/parameters/Id" }]
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/WarehouseRequest" }
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Warehouse" }
        "404": { $ref: "#/components/responses/Problem" }
  /admin/stock-items:
    get:
      tags: [inventory]
      operationId: listStockItems
      summary: List stock items with their total quantity on hand
      parameters:
        - name: type
          in: query
          schema: { $ref: "#/components/schemas/WarehouseType" }
        - name: q
          in: query
          description: Search in name or SKU
          schema: { type: string }
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema: { type: array, items: { $ref: "#/components/schemas/StockItem" } }
        "403": { $ref: "#/components/responses/Problem" }
    post:
      tags: [inventory]
      operationId: createStockItem
      summary: Create a stock item
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/StockItemRequest" }
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema: { $ref: "#/components/schemas/StockItem" }
        "422": { $ref: "#/components/responses/Problem" }
  /admin/stock-items/{id}:
    get:
      tags: [inventory]
      operationId: getStockItem
      summary: Get a stock item
      parameters: [{ $ref: "#/components/parameters/Id" }]
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema: { $ref: "#/components/schemas/StockItem" }
        "404": { $ref: "#/components/responses/Problem" }
    put:
      tags: [inventory]
      operationId: updateStockItem
      summary: Update a stock item and replace its packagings
      parameters: [{ $ref: "#/components/parameters/Id" }]
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/StockItemRequest" }
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema: { $ref: "#/components/schemas/StockItem" }
        "422": { $ref: "#/components/responses/Problem" }
  /admin/stock/balances:
    get:
      tags: [inventory]
      operationId: listStockBalances
      summary: Quantity and FIFO value per warehouse and item
      parameters:
        - name: warehouseId
          in: query
          schema: { type: string, format: uuid }
        - name: lowOnly
          in: query
          schema: { type: boolean, default: false }
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema: { type: array, items: { $ref: "#/components/schemas/StockBalance" } }
        "403": { $ref: "#/components/responses/Problem" }
  /admin/stock/movements:
    get:
      tags: [inventory]
      operationId: listStockMovements
      summary: Stock ledger, newest first
      parameters:
        - name: warehouseId
          in: query
          schema: { type: string, format: uuid }
        - name: stockItemId
          in: query
          schema: { type: string, format: uuid }
        - name: limit
          in: query
          schema: { type: integer, minimum: 1, maximum: 500, default: 100 }
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema: { type: array, items: { $ref: "#/components/schemas/StockMovement" } }
        "403": { $ref: "#/components/responses/Problem" }
  /admin/stock/lots:
    get:
      tags: [inventory]
      operationId: listStockLots
      summary: Open FIFO lots (oldest first)
      parameters:
        - name: warehouseId
          in: query
          schema: { type: string, format: uuid }
        - name: stockItemId
          in: query
          schema: { type: string, format: uuid }
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema: { type: array, items: { $ref: "#/components/schemas/StockLot" } }
        "403": { $ref: "#/components/responses/Problem" }
  /admin/stock/documents:
    get:
      tags: [inventory]
      operationId: listStockDocuments
      summary: Transfers, consumption notes, waste and inventory counts
      parameters:
        - name: type
          in: query
          schema: { $ref: "#/components/schemas/StockDocumentType" }
        - name: limit
          in: query
          schema: { type: integer, minimum: 1, maximum: 200, default: 50 }
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema: { type: array, items: { $ref: "#/components/schemas/StockDocument" } }
        "403": { $ref: "#/components/responses/Problem" }
    post:
      tags: [inventory]
      operationId: createStockDocument
      summary: Create and post a stock document (moves stock immediately)
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/StockDocumentRequest" }
      responses:
        "201":
          description: Posted
          content:
            application/json:
              schema: { $ref: "#/components/schemas/StockDocument" }
        "422": { $ref: "#/components/responses/Problem" }
  /admin/stock/documents/{id}:
    get:
      tags: [inventory]
      operationId: getStockDocument
      summary: Get a stock document
      parameters: [{ $ref: "#/components/parameters/Id" }]
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema: { $ref: "#/components/schemas/StockDocument" }
        "404": { $ref: "#/components/responses/Problem" }
  /admin/recipes/products/{productId}:
    get:
      tags: [inventory]
      operationId: getProductRecipe
      summary: Ingredients consumed when one unit of the product is sold
      parameters:
        - name: productId
          in: path
          required: true
          schema: { type: string, format: uuid }
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Recipe" }
        "403": { $ref: "#/components/responses/Problem" }
    put:
      tags: [inventory]
      operationId: replaceProductRecipe
      summary: Replace the product's recipe
      parameters:
        - name: productId
          in: path
          required: true
          schema: { type: string, format: uuid }
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/RecipeRequest" }
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Recipe" }
        "422": { $ref: "#/components/responses/Problem" }
  /admin/recipes/options/{optionId}:
    get:
      tags: [inventory]
      operationId: getOptionRecipe
      summary: Ingredients consumed by a modifier option (e.g. 30 g cheese sauce)
      parameters:
        - name: optionId
          in: path
          required: true
          schema: { type: string, format: uuid }
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Recipe" }
        "403": { $ref: "#/components/responses/Problem" }
    put:
      tags: [inventory]
      operationId: replaceOptionRecipe
      summary: Replace the modifier option's recipe
      parameters:
        - name: optionId
          in: path
          required: true
          schema: { type: string, format: uuid }
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/RecipeRequest" }
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Recipe" }
        "422": { $ref: "#/components/responses/Problem" }

  # ---------------------------------------------------------------- procurement
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  parameters:
...
    Id:
      name: id
      in: path
      required: true
      schema: { type: string, format: uuid }
...
  responses:
    Problem:
      description: Error (RFC 9457)
      content:
        application/problem+json:
          schema: { $ref: "#/components/schemas/Problem" }
  schemas:
...
    Problem:
      type: object
      properties:
        type: { type: string }
        title: { type: string }
        status: { type: integer }
        detail: { type: string }
        instance: { type: string }
        code: { type: string, description: Stable machine-readable error code }
        errors:
          type: array
          items:
            type: object
            properties:
              field: { type: string }
              message: { type: string }
    Money:
      type: object
      required: [amount, currency]
      properties:
        amount: { type: string, pattern: '^-?\d+(\.\d{1,2})?$', examples: ["20.00"] }
        currency: { type: string, examples: [RON] }
    LocalizedText:
      type: object
      description: Text in each supported locale; `ro` is required
      required: [ro]
      properties:
        ro: { type: string, minLength: 1 }
        en: { type: string }
...
    Quantity:
      type: string
      pattern: '^-?\d{1,9}(\.\d{1,3})?$'
      description: Decimal quantity, up to 3 decimals
      examples: ["2.500"]
    Dimension:
      type: string
      enum: [MASS, VOLUME, COUNT, LENGTH]
    UnitStatus:
      type: string
      enum: [ACTIVE, UNMAPPED]
    MeasureUnit:
      type: object
      required: [id, code, name, factor, isReference, aliases, status]
      properties:
        id: { type: string, format: uuid }
        code: { type: string, examples: [kg] }
        name: { $ref: "#/components/schemas/LocalizedText" }
        dimension: { $ref: "#/components/schemas/Dimension" }
        factor: { type: string, description: How many reference units one unit equals (kg = 1000) }
        isReference: { type: boolean }
        aliases: { type: array, items: { type: string } }
        status: { $ref: "#/components/schemas/UnitStatus" }
    MeasureUnitRequest:
      type: object
      required: [code, name, dimension, factor]
      properties:
        code: { type: string, minLength: 1, maxLength: 12 }
        name: { $ref: "#/components/schemas/LocalizedText" }
        dimension: { $ref: "#/components/schemas/Dimension" }
        factor: { type: string, pattern: '^\d{1,12}(\.\d{1,6})?$' }
        aliases: { type: array, items: { type: string, maxLength: 30 } }
    WarehouseType:
      type: string
      enum: [INGREDIENTS, FINISHED_GOODS, FIXED_ASSETS, CONSUMABLES]
    Warehouse:
      type: object
      required: [id, code, name, type, active]
      properties:
        id: { type: string, format: uuid }
        code: { type: string }
        name: { $ref: "#/components/schemas/LocalizedText" }
        type: { $ref: "#/components/schemas/WarehouseType" }
        active: { type: boolean }
    WarehouseRequest:
      type: object
      required: [code, name, type]
      properties:
        code: { type: string, minLength: 1, maxLength: 20, pattern: '^[A-Z0-9_]+$' }
        name: { $ref: "#/components/schemas/LocalizedText" }
        type: { $ref: "#/components/schemas/WarehouseType" }
        active: { type: boolean, default: true }
    Packaging:
      type: object
      required: [name, qtyInBase]
      properties:
        id: { type: string, format: uuid }
        name: { type: string, minLength: 1, maxLength: 40, examples: ["bax 24"] }
        qtyInBase: { $ref: "#/components/schemas/Quantity" }
        barcode: { type: string, maxLength: 40 }
    StockItem:
      type: object
      required: [id, sku, name, type, baseUnitId, baseUnitCode, minStock, active, packagings, quantityOnHand]
      properties:
        id: { type: string, format: uuid }
        sku: { type: string }
        name: { $ref: "#/components/schemas/LocalizedText" }
        type: { $ref: "#/components/schemas/WarehouseType" }
        baseUnitId: { type: string, format: uuid, description: "Reference unit of the item's dimension (g, ml, buc)" }
        baseUnitCode: { type: string }
        displayUnitId: { type: string, format: uuid }
        defaultWarehouseId: { type: string, format: uuid }
        minStock: { $ref: "#/components/schemas/Quantity" }
        active: { type: boolean }
        packagings: { type: array, items: { $ref: "#/components/schemas/Packaging" } }
        quantityOnHand: { $ref: "#/components/schemas/Quantity" }
    StockItemRequest:
      type: object
      required: [sku, name, type, dimension]
      properties:
        sku: { type: string, minLength: 1, maxLength: 40 }
        name: { $ref: "#/components/schemas/LocalizedText" }
        type: { $ref: "#/components/schemas/WarehouseType" }
        dimension: { $ref: "#/components/schemas/Dimension" }
        displayUnitId: { type: string, format: uuid }
        defaultWarehouseId: { type: string, format: uuid }
        minStock: { $ref: "#/components/schemas/Quantity" }
        active: { type: boolean, default: true }
        packagings: { type: array, items: { $ref: "#/components/schemas/Packaging" } }
    StockBalance:
      type: object
      required: [warehouseId, stockItemId, sku, itemName, quantity, unitCode, value, minStock, low]
      properties:
        warehouseId: { type: string, format: uuid }
        stockItemId: { type: string, format: uuid }
        sku: { type: string }
        itemName: { type: string }
        quantity: { $ref: "#/components/schemas/Quantity" }
        unitCode: { type: string }
        value: { $ref: "#/components/schemas/Money" }
        minStock: { $ref: "#/components/schemas/Quantity" }
        low: { type: boolean }
    StockMovementType:
      type: string
      enum: [RECEIPT_NIR, CONSUMPTION_SALE, RETURN_SALE, TRANSFER_OUT, TRANSFER_IN, CONSUMPTION_NOTE, WASTE, COUNT_LOSS, COUNT_GAIN, REVERSAL]
    StockMovement:
      type: object
      required: [id, warehouseId, stockItemId, itemName, type, quantity, unitCode, cost, occurredAt]
      properties:
        id: { type: string, format: uuid }
        warehouseId: { type: string, format: uuid }
        stockItemId: { type: string, format: uuid }
        itemName: { type: string }
        type: { $ref: "#/components/schemas/StockMovementType" }
        quantity: { $ref: "#/components/schemas/Quantity" }
        unitCode: { type: string }
        cost: { $ref: "#/components/schemas/Money" }
        sourceType: { type: string, examples: [NIR, ORDER, DOCUMENT] }
        sourceId: { type: string, format: uuid }
        sourceRef: { type: string, description: "Human reference, e.g. NIR-2026-000012 or order B-014" }
        occurredAt: { type: string, format: date-time }
    StockLot:
      type: object
      required: [id, warehouseId, stockItemId, receivedAt, quantityInitial, quantityRemaining, unitCost, unitCode]
      properties:
        id: { type: string, format: uuid }
        warehouseId: { type: string, format: uuid }
        stockItemId: { type: string, format: uuid }
        receivedAt: { type: string, format: date-time }
        quantityInitial: { $ref: "#/components/schemas/Quantity" }
        quantityRemaining: { $ref: "#/components/schemas/Quantity" }
        unitCost: { type: string, description: Cost per base unit (6 decimals) }
        unitCode: { type: string }
    StockDocumentType:
      type: string
      enum: [TRANSFER, CONSUMPTION, WASTE, COUNT]
    StockDocumentLineRequest:
      type: object
      required: [stockItemId, quantity]
      properties:
        stockItemId: { type: string, format: uuid }
        quantity: { $ref: "#/components/schemas/Quantity" }
        unitId: { type: string, format: uuid, description: Unit of `quantity`; defaults to the item's base unit }
        packagingId: { type: string, format: uuid, description: Alternative to unitId }
        reason: { type: string, maxLength: 200 }
    StockDocumentRequest:
      type: object
      required: [type, warehouseId, lines]
      properties:
        type: { $ref: "#/components/schemas/StockDocumentType" }
        date: { type: string, format: date }
        warehouseId: { type: string, format: uuid, description: Source warehouse (counted warehouse for COUNT) }
        targetWarehouseId: { type: string, format: uuid, description: Required for TRANSFER }
        note: { type: string, maxLength: 500 }
        lines:
          type: array
          minItems: 1
          description: For COUNT, `quantity` is the counted quantity
          items: { $ref: "#/components/schemas/StockDocumentLineRequest" }
    StockDocumentLine:
      type: object
      required: [stockItemId, itemName, quantity, unitCode]
      properties:
        stockItemId: { type: string, format: uuid }
        itemName: { type: string }
        quantity: { $ref: "#/components/schemas/Quantity" }
        unitCode: { type: string }
        systemQuantity: { $ref: "#/components/schemas/Quantity" }
        difference: { $ref: "#/components/schemas/Quantity" }
        cost: { $ref: "#/components/schemas/Money" }
        reason: { type: string }
    StockDocument:
      type: object
      required: [id, number, type, date, warehouseId, lines, totalCost, createdAt]
      properties:
        id: { type: string, format: uuid }
        number: { type: string, examples: ["TR-2026-000004"] }
        type: { $ref: "#/components/schemas/StockDocumentType" }
        date: { type: string, format: date }
        warehouseId: { type: string, format: uuid }
        targetWarehouseId: { type: string, format: uuid }
        note: { type: string }
        lines: { type: array, items: { $ref: "#/components/schemas/StockDocumentLine" } }
        totalCost: { $ref: "#/components/schemas/Money" }
        createdAt: { type: string, format: date-time }
    RecipeLineRequest:
      type: object
      required: [stockItemId, quantity]
      properties:
        stockItemId: { type: string, format: uuid }
        quantity: { $ref: "#/components/schemas/Quantity" }
        unitId: { type: string, format: uuid, description: Unit of `quantity`; defaults to the item's base unit }
        warehouseId: { type: string, format: uuid, description: Defaults to the item's default warehouse }
    RecipeRequest:
      type: object
      required: [lines]
      properties:
        lines: { type: array, items: { $ref: "#/components/schemas/RecipeLineRequest" } }
    RecipeLine:
      type: object
      required: [stockItemId, itemName, quantity, unitCode]
      properties:
        stockItemId: { type: string, format: uuid }
        itemName: { type: string }
        quantity: { $ref: "#/components/schemas/Quantity" }
        unitCode: { type: string, description: Base unit of the item }
        warehouseId: { type: string, format: uuid }
        estimatedCost: { $ref: "#/components/schemas/Money" }
    Recipe:
      type: object
      required: [lines]
      properties:
        productId: { type: string, format: uuid }
        optionId: { type: string, format: uuid }
        lines: { type: array, items: { $ref: "#/components/schemas/RecipeLine" } }
        estimatedCost: { $ref: "#/components/schemas/Money", description: Current FIFO cost of one portion }

    # ---------------------------------------------------------- procurement
...